Rename the crisp8 build product to crisp8.js.
VSCode expects node files to have either a .js extension or start with a node shebang. It has trouble understanding how to launch the built product without that.

Originally I intended to include a shebang in the node outputs so they could be run like regular binaries, but tsc always prepends referenced files before the shebang. Short of using a large build system (Gulp, Grunt, etc.) I don't plan to fix this any time soon. Without a shebang, the file can't be run directly on the shell (it must be passed as a parameter to the node binary) so we might as well give it a .js extension.