Many of the tests in the CRISP-8 project are going to follow the same format: a test binary blob fed into some system that generates a string debug dump of that system's internal state. For example, the main functional tests will be CRISP-8 ROMs that generate a snapshot of the CRISP-8 machine. The assembler will take assembly and output a ROM. The disassembler will take a ROM and output assembly.
Introduce the OutputTest, an abstract base class that represents this kind of test. It takes an input blob and an expected output string. Subclasses only need to override generateActualOutput() to take the input and transform it into the actual output string, for example, by running the input ROM in the virtual machine, or passing it through the assembler or disassembler.

A set of these OutputTests are run by the PlatformOutputTestRunner (Platform is a prefix that indicates the class should be implemented for a target platform, such as Node or the browser). Subclasses should implement loadNextTest which emits a concrete OutputTest each time its called, as well as the event-ish callbacks for reporting progress through the platform.